removed unncessary database init and dataaccess methods
