+++ src/main/java/com/aviva/DataAccess/AccountAccessInterface.java
@@ -1,5 +1,7 @@
 package com.aviva.DataAccess;
 
+import com.aviva.Entities.AccountHolder;
+
 import java.sql.ResultSet;
 
 /**
@@ -7,19 +9,6 @@ import java.sql.ResultSet;
  */
 
 public interface AccountAccessInterface {
-
-    /**
-     * Returns the ResultSet of the query to get a client's information from the database by their accountNumber
-     * @param accountNumber the account number of the client
-     * @return ResultSet of SQL query to get the client's information.
-     */
-    public ResultSet getClient(String accountNumber);
-
-    /**
-     * Returns the ResultSet of the query to get all clients' information from the database
-     * @return ResultSet of SQL query to get all clients' information.
-     */
-    public ResultSet getAllClients();
 
     /**
      * Returns the ResultSet of the query to get a particular client's latest credit score
@@ -34,4 +23,10 @@ public interface AccountAccessInterface {
      * @return ResultSet of SQL query to get all financial transactions stored for the client.
      */
     public ResultSet getAllFinancialTransactions(String accountNumber);
+
+    /**
+     * Inserts a car name against an account number into a table
+     * @param accountHolder the AccountHolder entity for which to insert the recommendations
+     */
+    public void insertRecommendedCars(AccountHolder accountHolder);
 }
