+++ src/main/java/com/aviva/DataAccess/CarAccessInterface.java
@@ -1,11 +1,6 @@
 package com.aviva.DataAccess;
 
-import com.aviva.Entities.Car;
-import com.aviva.Entities.Loan;
-
 import java.sql.ResultSet;
-import java.util.ArrayList;
-import java.util.HashMap;
 
 /**
  * This is the Data Access Interface for the data of the cars
@@ -26,10 +21,4 @@ public interface CarAccessInterface {
      */
     public ResultSet getAllCars();
 
-    /**
-     * Inserts a car name against an account number into a table
-     * @param accountNumber the account number of the client
-     * @param recommendations a map of Car Entities to their respective Loan entity
-     */
-    public void insertRecommendedCars(String accountNumber, HashMap<Car, Loan> recommendations);
 }
