+++ src/main/java/com/aviva/DataAccess/SQLCarDataAccess.java
@@ -1,13 +1,6 @@
 package com.aviva.DataAccess;
 
-import com.aviva.Entities.Car;
-import com.aviva.Entities.Installment;
-import com.aviva.Entities.Loan;
 import java.sql.*;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
 
 /**
  * Public class that handles querying data from the database for cars
@@ -71,129 +64,4 @@ public class SQLCarDataAccess implements CarAccessInterface {
         return null;
     }
 
-    /**
-     * Inserts a car name against an account number into a table
-     * @param accountNumber the account number of the client
-     * @param recommendations a map of Car Entities to their respective Loan entity
-     */
-    public void insertRecommendedCars(String accountNumber, HashMap<Car, Loan> recommendations) {
-        try {
-            // Establish connection with aviva database
-            Connection connection = DriverManager.getConnection(url, user, password);
-            Statement statement = connection.createStatement();
-            statement.execute("USE aviva");
-
-            // Query for existing recommendations for a client with the given account number
-            String command = "SELECT * FROM client WHERE accountnumber = '" + accountNumber + "'";
-            ResultSet rs = statement.executeQuery(command);
-
-            // Delete previous recommendations if 5 cars already recommended
-            deletePreviousRecommendations(connection, accountNumber);
-
-            if (rs.next()) {
-                // Loop through entry in the HashMap
-                for (Map.Entry<Car, Loan> entry : recommendations.entrySet()) {
-                    Car car = entry.getKey();
-                    Loan loan = entry.getValue();
-
-                    // Generate and set unique carID
-                    String uniqueID = UUID.randomUUID().toString();
-                    car.setID(uniqueID);
-
-                    insertIntoRecommendations(connection, accountNumber, car, loan);
-                    insertIntoInstallments(connection, loan, car.getID());
-                }
-            }
-            else {
-                System.out.println("Could not find specified account holder");
-            }
-        }
-        catch (SQLException e) {
-            System.out.println("Failed to insert recommendations into database");
-        }
-    }
-
-    /**
-     * Helper method to insert data into recommendations table
-     */
-    public void insertIntoRecommendations(Connection connection, String accountNumber, Car car, Loan loan) throws SQLException {
-        String command = "INSERT INTO recommendations (carID, accountNumber, carYear, carBrand, carMake, loanAmount, interestSum, capitalSum, loanSum, loanTerm, interestRate) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
-        PreparedStatement preparedStatement = connection.prepareStatement(command); // Prepare SQL Command
-
-        // Match variables against SQL data types
-        preparedStatement.setString(1, car.getID());
-        preparedStatement.setString(2, accountNumber);
-        preparedStatement.setInt(3, car.getYear());
-        preparedStatement.setString(4, car.getMake());
-        preparedStatement.setString(5, car.getModel());
-        preparedStatement.setFloat(6, loan.getLoanAmount());
-        preparedStatement.setFloat(7, loan.getInterestSum());
-        preparedStatement.setFloat(8, loan.getCapitalSum());
-        preparedStatement.setFloat(9, loan.getLoanSum());
-        preparedStatement.setInt(10, loan.getLoanTerm());
-        preparedStatement.setFloat(11, loan.getInterestRate());
-
-        // Execute SQL commands
-        preparedStatement.addBatch();
-        preparedStatement.executeBatch();
-    }
-
-    /**
-     * Helper method to insert data into installments table
-     */
-    public void insertIntoInstallments(Connection connection, Loan loan, String uniqueID) throws SQLException {
-        String command = "INSERT INTO recommendations (carID, termNumber, termCapital, termInterest, termInstallment, remainingAmount, interestSum) VALUES (?, ?, ?, ?, ?, ?, ?)";
-        PreparedStatement preparedStatement = connection.prepareStatement(command); // Prepare SQL Command
-
-        // Loop through each installment of the RecommendedCar
-        for (Installment installment: loan.getInstallments()) {
-            // Match variables against SQL data types
-            preparedStatement.setString(1, uniqueID);
-            preparedStatement.setInt(2, installment.getTermNumber());
-            preparedStatement.setFloat(3, installment.getTermCapital());
-            preparedStatement.setFloat(4, installment.getTermInterest());
-            preparedStatement.setFloat(5, installment.getTermInstallment());
-            preparedStatement.setFloat(6, installment.getRemainingAmount());
-            preparedStatement.setFloat(7, installment.getInterestSum());
-
-            preparedStatement.addBatch(); // Add to batch of statements to execute
-        }
-        preparedStatement.executeBatch(); // Execute SQL commands
-    }
-
-    /**
-     * Helper method to delete data from recommendations and installments tables if an account number already has 5 recommended cars
-     */
-    public void deletePreviousRecommendations (Connection connection, String accountNumber) {
-        ArrayList<String> carIDs = new ArrayList<>();
-
-        try {
-            // Query how many existing recommendations this account number has
-            Statement statement = connection.createStatement();
-            String command = "SELECT * FROM recommendations WHERE accountnumber = '" + accountNumber + "'";
-            ResultSet rs = statement.executeQuery(command);
-
-            // Add IDs of existing recommendations to a list
-            while (rs.next()) {
-                carIDs.add(rs.getString(1));
-                }
-
-            // Delete all records of recommendations and installments if there are already 5 recommendations
-            if (carIDs.size() >= 5) {
-                // Delete all records from recommendations table
-                command = "DELETE FROM recommendations WHERE accountnumber = '" + accountNumber + "'";
-                statement.execute(command);
-
-                // Delete all records from installments table
-                for (String ID : carIDs) {
-                    command = "DELETE FROM installments WHERE carID = '" + ID + "'";
-                    statement.execute(command);
-                }
-            }
-        }
-        catch (SQLException e) {
-            System.out.println("This account number has less than 5 recommended cars");
-        }
-
-    }
 }
